implemented send again for user messages when user message not received the bot and few MarkDown issues

+++ BotsSDK/korebotsdklib/src/main/java/kore/botssdk/net/RestResponse.java
@@ -1,8 +1,6 @@
 package kore.botssdk.net;
 
 
-import androidx.annotation.Nullable;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 
@@ -178,7 +176,7 @@ public class RestResponse {
         private BotMessage message;
         private String resourceid = "/bot.message";
         private BotInfoModel botInfo;
-        private double clientMessageId = System.currentTimeMillis();
+        private long clientMessageId = System.currentTimeMillis();
         private Meta meta;
         private Object id = clientMessageId;
         private String client = "Android";
@@ -204,7 +202,7 @@ public class RestResponse {
             this.botInfo = botInfo;
         }
 
-        public double getClientMessageId() {
+        public long getClientMessageId() {
             return clientMessageId;
         }
